Checkup, vmi: Move custom checkup code
Currently, both VMI under test and traffic-gen are built by
the shared `NewDPDKVMI` function.

In an effort to move all checkup-specific logic away from `spec.go`,
Move the `NewDPDKVMI` function and its accompanying DPDKVMIConfig
struct, to the `vmi.go` file in the `checkup` package.

Also move the `CloudInit` and `RandomizeName` functions.
